requirements.txt
# Webフレームワーク & UI

gradio==5.42.0

fastapi>=0.110.0

pydantic>=2.0.0

annotated-types>=0.4.0

uvicorn

 

# LangChain関連: RAGのフレームワーク

langchain==0.3.27

langchain-openai==0.3.31

langchain-community==0.3.27

langchain-chroma==0.2.5

 

# ベクトルデータベース

chromadb==1.0.20

 

# OpenAI API関連

openai

tiktoken==0.11.0

 

# 設定ファイル(.env)の読み込み

python-dotenv==1.1.1

 

# データハンドリング

pandas

numpy

 

# その他

requests==2.32.4
spotipy==2.24.0